sources/prod/requirements.txt
kafka-python~=2.1.5